chore: Adapt Theme.kt for CMP
